+++ src/pages/favorites/Index.tsx
